requirements.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
future==0.16.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pylint==1.9.1
python-twitter==3.4.1
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
